Feature 부스 북마크 api (#20)
* feat: add controller, DTO, entity, repository, and service layers

* Feat:booth apis

* Feat: like api

* Feat: boothBookmark api

---------

+++ src/main/java/com/jnu/festival/domain/like/dto/LikeRequestDTO.java
@@ -6,11 +6,8 @@ import lombok.Setter;
 @Getter
 @Setter
 public class LikeRequestDTO {
-
-    private int userId;
-    private int boothId;
+    private Long userId;
+    private Long boothId;
     private boolean isDeleted;
 
 }
-
-
